Make loggable resources specify their log presenter
Instead of magically finding it
