+++ decidim-core/app/models/decidim/participatory_process_user_role.rb
@@ -5,6 +5,7 @@ module Decidim
   # kind of relation does the user has.
   class ParticipatoryProcessUserRole < ApplicationRecord
     include Traceable
+    include Loggable
 
     belongs_to :user, foreign_key: "decidim_user_id", class_name: "Decidim::User", optional: true
     belongs_to :participatory_process, foreign_key: "decidim_participatory_process_id", class_name: "Decidim::ParticipatoryProcess", optional: true
@@ -14,6 +15,10 @@ module Decidim
     ROLES = %w(admin collaborator moderator).freeze
     validates :role, inclusion: { in: ROLES }, uniqueness: { scope: [:user, :participatory_process] }
     validate :user_and_participatory_process_same_organization
+
+    def self.log_presenter_class_for(_log)
+      Decidim::ParticipatoryProcesses::AdminLog::ParticipatoryProcessUserRolePresenter
+    end
 
     private
