+++ decidim-core/spec/models/decidim/feature_spec.rb
@@ -12,5 +12,10 @@ module Decidim
     it { is_expected.to be_versioned }
 
     include_examples "publicable"
+
+    it "overwrites the log presenter" do
+      expect(described_class.log_presenter_class_for(:foo))
+        .to eq Decidim::AdminLog::FeaturePresenter
+    end
   end
 end
