+++ decidim-core/spec/models/decidim/moderation_spec.rb
@@ -10,5 +10,10 @@ module Decidim
 
     it { is_expected.to be_valid }
     it { is_expected.to be_versioned }
+
+    it "overwrites the log presenter" do
+      expect(described_class.log_presenter_class_for(:foo))
+        .to eq Decidim::AdminLog::ModerationPresenter
+    end
   end
 end
