+++ decidim-core/spec/models/decidim/organization_spec.rb
@@ -9,6 +9,11 @@ module Decidim
     it { is_expected.to be_valid }
     it { is_expected.to be_versioned }
 
+    it "overwrites the log presenter" do
+      expect(described_class.log_presenter_class_for(:foo))
+        .to eq Decidim::AdminLog::OrganizationPresenter
+    end
+
     describe "has an association for scopes" do
       subject(:organization_scopes) { organization.scopes }
