+++ decidim-core/spec/models/decidim/static_page_spec.rb
@@ -8,6 +8,11 @@ module Decidim
 
     it { is_expected.to be_versioned }
 
+    it "overwrites the log presenter" do
+      expect(described_class.log_presenter_class_for(:foo))
+        .to eq Decidim::AdminLog::StaticPagePresenter
+    end
+
     describe "validations" do
       let(:invalid_slug) { "#Invalid.Slug" }
