+++ decidim-core/spec/models/decidim/user_group_spec.rb
@@ -10,6 +10,11 @@ module Decidim
 
     it { is_expected.to be_versioned }
 
+    it "overwrites the log presenter" do
+      expect(described_class.log_presenter_class_for(:foo))
+        .to eq Decidim::AdminLog::UserGroupPresenter
+    end
+
     it "is valid" do
       expect(subject).to be_valid
     end
