+++ decidim-participatory_processes/app/models/decidim/participatory_process.rb
@@ -15,6 +15,7 @@ module Decidim
     include Decidim::Followable
     include Decidim::HasReference
     include Decidim::Traceable
+    include Decidim::Loggable
 
     belongs_to :organization,
                foreign_key: "decidim_organization_id",
@@ -61,6 +62,10 @@ module Decidim
     # Returns an ActiveRecord::Relation.
     def self.promoted
       where(promoted: true)
+    end
+
+    def self.log_presenter_class_for(_log)
+      Decidim::ParticipatoryProcesses::AdminLog::ParticipatoryProcessPresenter
     end
 
     def hashtag
