+++ decidim-participatory_processes/app/presenters/decidim/participatory_processes/admin_log/participatory_process_presenter.rb
@@ -1,0 +1,56 @@
+# frozen_string_literal: true
+
+module Decidim
+  module ParticipatoryProcesses
+    module AdminLog
+      # This class holds the logic to present a `Decidim::ParticipatoryProcess`
+      # for the `AdminLog` log.
+      #
+      # Usage should be automatic and you shouldn't need to call this class
+      # directly, but here's an example:
+      #
+      #    action_log = Decidim::ActionLog.last
+      #    view_helpers # => this comes from the views
+      #    ParticipatoryProcessPresenter.new(action_log, view_helpers).present
+      class ParticipatoryProcessPresenter < Decidim::Log::BasePresenter
+        private
+
+        def diff_fields_mapping
+          {
+            description: :i18n,
+            developer_group: :i18n,
+            hashtag: :string,
+            local_area: :i18n,
+            meta_scope: :i18n,
+            participatory_scope: :i18n,
+            participatory_structure: :i18n,
+            published_at: :date,
+            scopes_enabled: :boolean,
+            short_description: :i18n,
+            slug: :string,
+            subtitle: :i18n,
+            target: :i18n,
+            title: :i18n
+          }
+        end
+
+        def action_string
+          case action
+          when "create", "publish", "unpublish"
+            "decidim.admin_log.participatory_process.#{action}"
+          else
+            super
+          end
+        end
+
+        def i18n_labels_scope
+          "activemodel.attributes.participatory_process"
+        end
+
+        def has_diff?
+          action == "unpublish" || super
+        end
+      end
+    end
+  end
+end
