+++ decidim-participatory_processes/spec/models/decidim/participatory_process_spec.rb
@@ -14,6 +14,11 @@ module Decidim
 
     include_examples "publicable"
 
+    it "overwrites the log presenter" do
+      expect(described_class.log_presenter_class_for(:foo))
+        .to eq Decidim::ParticipatoryProcesses::AdminLog::ParticipatoryProcessPresenter
+    end
+
     context "when there's a process with the same slug in the same organization" do
       let!(:external_process) { create :participatory_process, organization: participatory_process.organization, slug: "my-slug" }
